accidentally worked on weighted rr. oops!
