+++ pox/misc/loadbalancing/wlc_rand.py
@@ -1,0 +1,3 @@
+from pox.misc.loadbalancing.lb_weighted_least_connection import WeightedLeastConnection
+
+class RandWLC(WeightedLeastConnection):
